* commit '78d61984610328485ca8dbec5cd56f13f456037e':
  NMAAS-302: Set missing dependency
  NMAAS-302: Used dedicated class to prepare helm install variables for ingress
  NMAAS-302: Moved ingress control logic to KubernetesManager
  NMAAS-302: Use KClusterIngress property to control ingress resource handling
  NMAAS-302: Updated default test domain name to be in lowercase
  NMAAS-302: Modified ingress resource manager interface
  NMAAS-302: Implemented option to not configure ingress through k8s api

+++ nmaas-platform/src/main/java/net/geant/nmaas/nmservice/deployment/containerorchestrators/kubernetes/KubernetesManager.java
@@ -1,5 +1,6 @@
 package net.geant.nmaas.nmservice.deployment.containerorchestrators.kubernetes;
 
+import net.geant.nmaas.externalservices.inventory.kubernetes.KClusterIngressManager;
 import net.geant.nmaas.nmservice.deployment.ContainerOrchestrator;
 import net.geant.nmaas.nmservice.deployment.containerorchestrators.kubernetes.components.cluster.KClusterCheckException;
 import net.geant.nmaas.nmservice.deployment.containerorchestrators.kubernetes.components.ingress.IngressControllerManipulationException;
@@ -32,6 +33,7 @@ public class KubernetesManager implements ContainerOrchestrator {
     private KClusterValidator clusterValidator;
     private KServiceLifecycleManager serviceLifecycleManager;
     private KServiceOperationsManager serviceOperationsManager;
+    private KClusterIngressManager clusterIngressManager;
     private IngressControllerManager ingressControllerManager;
     private IngressResourceManager ingressResourceManager;
 
@@ -40,12 +42,14 @@ public class KubernetesManager implements ContainerOrchestrator {
                              KClusterValidator clusterValidator,
                              KServiceLifecycleManager serviceLifecycleManager,
                              KServiceOperationsManager serviceOperationsManager,
+                             KClusterIngressManager clusterIngressManager,
                              IngressControllerManager ingressControllerManager,
                              IngressResourceManager ingressResourceManager) {
         this.repositoryManager = repositoryManager;
         this.clusterValidator = clusterValidator;
         this.serviceLifecycleManager = serviceLifecycleManager;
         this.serviceOperationsManager = serviceOperationsManager;
+        this.clusterIngressManager = clusterIngressManager;
         this.ingressControllerManager = ingressControllerManager;
         this.ingressResourceManager = ingressResourceManager;
     }
@@ -81,8 +85,10 @@ public class KubernetesManager implements ContainerOrchestrator {
     public void prepareDeploymentEnvironment(Identifier deploymentId)
             throws CouldNotPrepareEnvironmentException, ContainerOrchestratorInternalErrorException {
         try {
-            String domain = repositoryManager.loadDomain(deploymentId);
-            ingressControllerManager.deployIngressControllerIfMissing(domain);
+            if(!clusterIngressManager.getUseExistingController()) {
+                String domain = repositoryManager.loadDomain(deploymentId);
+                ingressControllerManager.deployIngressControllerIfMissing(domain);
+            }
         } catch (InvalidDeploymentIdException idie) {
             throw new ContainerOrchestratorInternalErrorException(
                     "Service not found in repository -> Invalid deployment id " + idie.getMessage());
@@ -97,12 +103,18 @@ public class KubernetesManager implements ContainerOrchestrator {
             throws CouldNotDeployNmServiceException, ContainerOrchestratorInternalErrorException {
         try {
             KubernetesNmServiceInfo service = repositoryManager.loadService(deploymentId);
-            serviceLifecycleManager.deployService(deploymentId);
-            String serviceExternalUrl = ingressResourceManager.createOrUpdateIngressResource(
-                    deploymentId,
+            String serviceExternalUrl = ingressResourceManager.generateServiceExternalURL(
                     service.getDomain(),
-                    service.getDeploymentName());
+                    service.getDeploymentName(),
+                    clusterIngressManager.getExternalServiceDomain());
             repositoryManager.updateKServiceExternalUrl(deploymentId, serviceExternalUrl);
+            serviceLifecycleManager.deployService(deploymentId);
+            if (!clusterIngressManager.getUseExistingIngress()) {
+                ingressResourceManager.createOrUpdateIngressResource(
+                        deploymentId,
+                        service.getDomain(),
+                        serviceExternalUrl);
+            }
         } catch (InvalidDeploymentIdException idie) {
             throw new ContainerOrchestratorInternalErrorException(
                     "Service not found in repository -> Invalid deployment id " + idie.getMessage());
@@ -129,7 +141,9 @@ public class KubernetesManager implements ContainerOrchestrator {
         try {
             serviceLifecycleManager.deleteService(deploymentId);
             KubernetesNmServiceInfo service = repositoryManager.loadService(deploymentId);
-            ingressResourceManager.deleteIngressRule(service.getServiceExternalUrl(), service.getDomain());
+            if (!clusterIngressManager.getUseExistingIngress()) {
+                ingressResourceManager.deleteIngressRule(service.getServiceExternalUrl(), service.getDomain());
+            }
         } catch (InvalidDeploymentIdException idie) {
             throw new ContainerOrchestratorInternalErrorException(
                     "Service not found in repository -> Invalid deployment id " + idie.getMessage());
